refactor(core): move interaction value types + TitleConfig into leviath-core
Phase 5a of the engine relocation: move the plain serde value types the run
engine passes across the InteractionBackend seam (InteractionRequest,
InteractionResponse, InteractionKind, ApprovalScope, BodyFormat + pure
helpers) into leviath_core::interaction, and TitleConfig into
leviath_core::config. Both are re-exported from cli (interaction.rs /
config.rs) so existing crate::interaction::* / crate::config::TitleConfig
paths resolve unchanged. This lets leviath-runtime reference these types in
the next sub-steps without depending on leviath-cli.
